Исправление по замечаниям PR #4 -Разбиение монолитного JsonLoad на классы dto и сущность JsonLoader по парсингу данных.
