+++ app/src/main/java/com/alohagoha/aaaa/MovieApp.kt
@@ -1,0 +1,8 @@
+package com.alohagoha.aaaa
+
+import android.app.Application
+import com.alohagoha.aaaa.data.JsonLoader
+
+class MovieApp : Application() {
+    val jsonLoader: JsonLoader by lazy { JsonLoader(applicationContext) }
+}
