+++ app/src/main/java/com/alohagoha/aaaa/data/dto/JsonActor.kt
@@ -1,0 +1,12 @@
+package com.alohagoha.aaaa.data.dto
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class JsonActor(
+    val id: Int,
+    val name: String,
+    @SerialName("profile_path")
+    val profilePicture: String
+)
