+++ app/src/main/java/com/alohagoha/aaaa/data/dto/JsonGenre.kt
@@ -1,0 +1,6 @@
+package com.alohagoha.aaaa.data.dto
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class JsonGenre(val id: Int, val name: String)
