+++ app/src/main/java/com/alohagoha/aaaa/data/dto/JsonMovie.kt
@@ -1,0 +1,24 @@
+package com.alohagoha.aaaa.data.dto
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class JsonMovie(
+    val id: Int,
+    val title: String,
+    @SerialName("poster_path")
+    val posterPicture: String,
+    @SerialName("backdrop_path")
+    val backdropPicture: String,
+    val runtime: Int,
+    @SerialName("genre_ids")
+    val genreIds: List<Int>,
+    val actors: List<Int>,
+    @SerialName("vote_average")
+    val ratings: Float,
+    @SerialName("vote_count")
+    val votesCount: Int,
+    val overview: String,
+    val adult: Boolean
+)
